可用积分增加的修复，空间人员数量的修复

+++ src/main/java/com/xin/graphdomainbackend/model/vo/SpaceVO.java
@@ -57,7 +57,7 @@ public class SpaceVO implements Serializable {
     private UserVO user;
 
     // 空间成员数量
-    private Long memberCount;
+    private Integer memberCount;
 
     /**
      * 权限列表
